finalize fde implementations

+++ src/meow/fde/lumerical.py
@@ -1,7 +1,11 @@
-""" FDE Lumerical Backend """
+"""FDE Lumerical Backend."""
+
+from typing import Any
 
 import numpy as np
-from pydantic.v1.types import PositiveInt
+from pydantic import PositiveInt
+
+from meow.arrays import ComplexArray2D
 
 from ..cross_section import CrossSection
 from ..environment import Environment
@@ -9,25 +13,33 @@ from ..mode import Mode, normalize_product, zero_phase
 from ..structures import Structure3D
 
 _global = {"sim": None}
+Sim = Any
 
 
-def get_sim(**kwargs):
+def get_sim(**kwargs: Any) -> Sim:
+    """Get the Lumerical simulation object."""
     sim = kwargs.get("sim", None)
     if sim is not None:
         return sim
     sim = _global["sim"]
     if sim is None:
-        raise ValueError(
-            "Could not start Lumerical simulation. "
-            "Please either pass the `lumapi.MODE` simulation object as an argument to `compute_modes_lumerical` or "
-            "use `set_sim(sim)` to globally set the lumapi.MODE simulation object."
+        msg = (
+            "Could not start Lumerical simulation. Please either pass the "
+            "`lumapi.MODE` simulation object as an argument to "
+            "`compute_modes_lumerical` or use `set_sim(sim)` to globally "
+            "set the lumapi.MODE simulation object."
         )
+        raise ValueError(msg)
     return sim
 
 
 def create_lumerical_geometries(
-    sim, structures: list[Structure3D], env: Environment, unit: float
-):
+    sim: Sim,
+    structures: list[Structure3D],
+    env: Environment,
+    unit: float,
+) -> None:
+    """Create Lumerical geometries from a list of structures."""
     sim = get_sim(sim=sim)
     sim.switchtolayout()
     sim.deleteall()
@@ -35,32 +47,36 @@ def create_lumerical_geometries(
         s._lumadd(sim, env, unit, "yzx")
 
 
-# @validate_arguments
 def compute_modes_lumerical(
     cs: CrossSection,
     num_modes: PositiveInt = 10,
-    sim=None,
+    sim: Sim | None = None,
     unit: float = 1e-6,
-):
-    """compute ``Modes` for a given ``FdeSpec` (Lumerical backend)
-
-    Args:
-        sim: the lumerical simulation object
-        spec: The FDE simulation specification
-        unit: Conversion factor between MEOW unit (probably um) and Lumerical unit (probably m).
-    """
-    from lumapi import LumApiError  # fmt: skip # type: ignore
+) -> list[Mode]:
+    """Compute ``Modes` for a given ``FdeSpec` (Lumerical backend)."""
+    from lumapi import LumApiError  # type: ignore[reportMissingImports]
 
     sim = get_sim(sim=sim)
+
+    if sim is None:
+        msg = "Lumerical simulation object not given or found."
+        raise RuntimeError(msg)
+
     cell = cs._cell
-    assert cell is not None
+
+    if cell is None:
+        msg = (
+            "The cross-section does not have a cell defined. "
+            "Please define a cell before computing modes."
+        )
+        raise ValueError(msg)
 
     _assert_default_mesh_setting(cell.mesh.angle_phi == 0, "angle_phi")
     _assert_default_mesh_setting(cell.mesh.angle_theta == 0, "angle_theta")
     _assert_default_mesh_setting(cell.mesh.bend_radius is None, "bend_radius")
 
-    assert sim is not None
     create_lumerical_geometries(sim, cell.structures, cs.env, unit)
+
     sim.select("FDE")
     sim.delete()
     pml_settings = {}
@@ -102,7 +118,7 @@ def compute_modes_lumerical(
         sim.setnamed("FDE", "mesh cells z", cell.mesh.y_.shape[0] - num_pml_z)
     sim.setanalysis("number of trial modes", int(num_modes))
     sim.setanalysis("search", "near n")
-    sim.setanalysis("use max index", True)
+    sim.setanalysis("use max index", True)  # noqa: FBT003
     sim.setanalysis("wavelength", float(cs.env.wl * unit))
     sim.findmodes()
     modes = []
@@ -127,7 +143,16 @@ def compute_modes_lumerical(
     return sorted(modes, key=lambda m: np.real(m.neff), reverse=True)
 
 
-def _lumerical_fields_to_mode(cs, lneff, lEx, lEy, lEz, lHx, lHy, lHz):
+def _lumerical_fields_to_mode(
+    cs: CrossSection,
+    lneff: ComplexArray2D,
+    lEx: ComplexArray2D,
+    lEy: ComplexArray2D,
+    lEz: ComplexArray2D,
+    lHx: ComplexArray2D,
+    lHy: ComplexArray2D,
+    lHz: ComplexArray2D,
+) -> Mode:
     return Mode(
         cs=cs,
         neff=lneff.ravel().item(),
@@ -140,14 +165,16 @@ def _lumerical_fields_to_mode(cs, lneff, lEx, lEy, lEz, lHx, lHy, lHz):
     )
 
 
-def _assert_default_mesh_setting(condition, param_name):
+def _assert_default_mesh_setting(condition: bool, param_name: str) -> None:  # noqa: FBT001
     if not condition:
-        raise NotImplementedError(
-            f"Setting mesh.{param_name} is currently not supported in the Lumerical Backend. "
-            "Please open an issue of submit a PR on GitHub to fix this: ",
-            "https://github.com/flaport/meow",
+        msg = (
+            f"Setting mesh.{param_name} is currently not supported in the "
+            "Lumerical Backend. Please open an issue of submit a PR on GitHub "
+            "to fix this: https://github.com/flaport/meow",
         )
+        raise NotImplementedError(msg)
 
 
-def set_sim(sim):
+def set_sim(sim: Sim) -> None:
+    """Sets the global Lumerical simulation object."""
     _global["sim"] = sim
